# docs/reference/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Ceres engine's HTTP API.",
    "title": "Ceres",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/components": {
      "get": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Describe every component the caller may view.",
        "tags": [
          "components"
        ]
      }
    },
...
    "/api/components/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Describe one component and its children.",
        "tags": [
          "components"
        ]
      }
    },
    "/api/components/{id}/actions": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List one component's actions.",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/components/{id}/actions/{second}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one action binding.",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/components/{id}/actions/{second}/call": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Call an action with body arguments, access checked by the operation.",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/components/{id}/config": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one component's configuration.",
        "tags": [
          "components"
        ]
      }
    },
    "/api/components/{id}/connections": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List one component's connections with their connectivity.",
        "tags": [
          "components"
        ]
      }
    },
    "/api/components/{id}/connections/{second}/send": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send data over one component connection.",
        "tags": [
          "components"
        ]
      }
    },
    "/api/components/{id}/jobs": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List one component's scheduled jobs.",
        "tags": [
          "components"
        ]
      }
    },
    "/api/components/{id}/procedures": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List one component's procedures.",
        "tags": [
          "procedures"
        ]
      }
    },
    "/api/components/{id}/procedures/{second}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one procedure binding.",
        "tags": [
          "procedures"
        ]
      }
    },
    "/api/components/{id}/procedures/{second}/call": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Call a procedure with query arguments, access checked by the operation.",
        "tags": [
          "procedures"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Call a procedure with body arguments, access checked by the operation.",
        "tags": [
          "procedures"
        ]
      }
    },
    "/api/components/{id}/queries": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List one component's queries.",
        "tags": [
          "queries"
        ]
      }
    },
    "/api/components/{id}/queries/{second}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one query binding.",
        "tags": [
          "queries"
        ]
      }
    },
    "/api/components/{id}/queries/{second}/call": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Call a query with query arguments, access checked by the operation.",
        "tags": [
          "queries"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Call a query with body arguments, access checked by the operation.",
        "tags": [
          "queries"
        ]
      }
    },
...
    "/api/disable": {
      "post": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable matching components.",
        "tags": [
          "engine"
        ]
      }
    },
    "/api/down": {
      "post": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable and stop matching components.",
        "tags": [
          "engine"
        ]
      }
    },
    "/api/enable": {
      "post": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable matching components.",
        "tags": [
          "engine"
        ]
      }
    },
    "/api/groups": {
      "get": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List groups.",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a group.",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/count": {
      "get": {
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Count groups.",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/groups/{id}/members/{second}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "in": "path",
            "name": "second",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The request succeeded."
          },
          "default": {
            "description": "The request failed, carrying an error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a member from a group.",
        "tags": [
          "groups"
        ]
      }
    },
...
  },
...
}